requirements.txt
accelerate==0.27.2
benepar==0.2.0
bitarray==2.9.2
bitsandbytes==0.43.0
datasets==2.18.0
einops==0.7.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl#sha256=1618fc62e7aedc2b778caa3ea6e9c8280beac3b83fe7b93316cbb9e8f3abe7c5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
evaluate==0.4.1
fairscale==0.4.0
filelock==3.13.1
flash-attn==2.4.2
gradio==3.37.0
gradio_client==0.8.0
grpcio==1.62.1
h11==0.14.0
httpcore==0.17.3
httpx==0.25.1
huggingface-hub==0.20.2
ImageHash==4.3.1
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==2.2.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python-headless==4.8.0.74
pandas==2.1.4
peft==0.9.0
pillow==10.2.0
pipx==1.5.0
prettytable==3.9.0
progressbar==2.5
protobuf==3.20.3
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
PyYAML==6.0.1
regex==2023.12.25
safetensors==0.4.2
scikit-learn==1.5.1
scipy==1.8.0
seaborn==0.13.2
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==1.8
timm==0.4.12
tokenizers==0.15.1
torch==1.13.1
torch-struct==0.5
torchvision==0.14.0
tqdm==4.66.1
transformers==4.37.1
wandb==0.16.4
xformers==0.0.16
rapidfuzz==3.6.1
gensim==4.3.2
sentence-transformers==2.7.0
Pattern==3.6